[issue988] Get rid of LandmarkNodeComparer struct.
It was used only once in the LandmarkGraph::dump(). I don't see a
justification for why landmarks should be sorted as defined in the
comparer (the hierarchy was "less facts first" and "smaller fact
values" for tie-breaking). Hence, we might as well remove it.
Without the comparer, landmarks are dumped in order of their ID.