[issue1079] translator: clean-up (changing output) and type annotations
For users:

The change leads in a few domains to a different FDR encoding.

For developers:

We filter unreachable atoms from mutex groups at an earlier stage than
before.  We also observed that static facts have been removed from goals
before the translation to FDR. This FDR translation previously handled
these cases specially, which is no longer necessary. In a few domains this
leads to a different FDR encoding.

We also added some type annotations, in particular to the core data
structures.