[issue1076] Replace OSI with our own interface to LP/MIP solvers
For LP/MIP users: so far, we used the open solver interface OSI to communicate
with MIP and LP solvers. We now replaced this with our own interface to have
more direct control and fewer external dependencies. We now support CPLEX
22.11 and SoPlex from their current main branch (released versions <= 6.0.3 are
incompatible with C++-20). We dropped command line options for CLP and Gurobi
in the process: CLP has much worse performance than SoPlex which is also open
source, while Gurobi was never practically possible to compile with, as we did
not have a CMake find script for it. Performance with CPLEX increased in MIP
configurations and stayed roughly the same in LP configurations. Performance
with SoPlex increased for potential heuristics and decreased slightly in other
configurations.