[trivial] Fix timer computation on Mac.
The previous commit triggered a compile error on Mac due to implicit
conversions from floating-point to integer. It turns out the previous
rounding happened at a problematic state and could lead to very wrong
timing information on some (non-Intel) Mac platforms. For example, see
https://stackoverflow.com/questions/23378063/how-can-i-use-mach-absolute-time-without-overflowing

We changed the way these times are computed to avoid rounding at this
stage and be less convoluted.