[issue454] Move computation of negated axioms to search component
Negated axioms are now only computed for heuristics that need it. Specifically these are:
- additive heuristic
- context-enhanced additive heuristic
- causal graph heuristic
- FF heuristic
- max heuristic
- landmark heuristics (landmark sum and landmark cost partitioning)

Each heuristic computes the negated axioms separately. This can negatively
affect performance if a search with several such heuristics is performed, as is
for example done in LAMA (ff and landmark sum).
Heuristics that support axioms but do not need negated axioms might now be 
faster tasks with axioms, since they avoid the unnecessary computation.

A new enum option "axioms" has been added to all heuristics that need negated
axioms to determine how the negated axioms are computed.
- overapproximate_negated_cycles (default): Computes exact negation except for
  derived variables that have cyclic dependencies, which are trivially
  overapproximated (see issue453). This is the old behavior that can lead to an
  exponential explosion.
- overapproximate_negated: Trivially overapproximates all negations. This
  avoids the exponential explosion but makes the heuristic weaker.