[issue1173] Fix parameter order for MergeAndShrinkAlgorithm.
We passed the options for the merge_and_shrink evaluator into
the constructor of an internal class in the wrong order, leading to
wrong behaviour. For example, the boolean option prune_unreachable_states
was interpreted as max_states, so that max_states was set to 0
or 1 based on the boolean flag. This is now fixed.