[issue1074] Change behavior of option to compute only causal landmarks.
We remove the option to consider only causal landmarks (`only_causal_landmarks`)
from the landmark factories `lm_exhaust` and `lm_rhw`. For `lm_exhaust` we add a
new boolean option `use_unary_relaxation` which yields an equivalent landmark
graph as a result. Its effect in the landmark computation is different, though,
as it no longer constitutes a post-processing to remove all non-causal
landmarks, but makes sure only causal landmarks are added in the first place.
This results in lower landmark generation time for `lm_exhaust` with this option
active; on average, landmark generation takes 25-30% of the time it used to.

Removing the option from `lm_rhw` affects the portfolios seq-sat-fdss-2018 and
seq-sat-fdss-2023, which used to have some configurations with the flag on and
others with the flag off. Simply removing the option (essentially turning the
flag off in all configurations) didn't have a negative effect as can be seen
here:
https://ai.dmi.unibas.ch/_experiments/ai/downward/issue1074/data/issue1074-v5-sat-eval/issue1074-v5-portfolios-issue1074-base-issue1074-v5-compare.html